Implements email handler and improvements in several functionalities

+++ preprocessor/helpers/statistics.py
@@ -4,11 +4,11 @@ from typing import List, Dict
 def compute_statistics(data_list: List[float]) -> Dict[str, float]:
     data_array = np.array(data_list)
     return {
-        "mean": np.mean(data_array),
-        "median": np.median(data_array),
-        "std_dev": np.std(data_array),
-        "min": np.min(data_array),
-        "max": np.max(data_array),
-        "range": np.ptp(data_array),
+        "mean": float(np.mean(data_array)),
+        "median": float(np.median(data_array)),
+        "std_dev": float(np.std(data_array)),
+        "min": float(np.min(data_array)),
+        "max": float(np.max(data_array)),
+        "range": float(np.ptp(data_array)),
     }
